+++ src/vs/platform/externalServices/common/serviceMachineId.ts
